src/requirements.txt
sql-metadata
argparse
duckdb
pyarrow
Flask
gunicorn
google-cloud-storage
pandas